Lesson 26 (Optional)
Was about to start doing the optional buzzing/virbation stuff, but can't be bothered to hook up my actual device. no real need to I think as I understand ewhats going on with LiveData -> ViewModel -> View etc etc to get Mvvm setup. Not quite as nice as Xamarin+MvvmCross but its not that different either.